itoa subroutine implemented
i did not include making sure it fits into space it has as in case of this particular case this will never be a problem because we will be printing numbers less than or equal to the one given by user in the input step and we will use the same memory for creating strings for numbers. leq numbers = leq digits = leq characters. good luck everyone who will borrow this code for something important without reading this message